+++ ui-lib/src/androidTest/java/co/electriccoin/zcash/ui/fixture/MockSynchronizer.kt
@@ -207,7 +207,7 @@ internal class MockSynchronizer : CloseableSynchronizer {
         error("Intentionally not implemented in ${MockSynchronizer::class.simpleName} implementation.")
     }
 
-    override suspend fun getWalletDbPath(): String {
+    override suspend fun getWalletDbPathForVoting(): String {
         error("Intentionally not implemented in ${MockSynchronizer::class.simpleName} implementation.")
     }