+++ ui-lib/src/main/java/co/electriccoin/zcash/ui/common/usecase/CreateFlexaTransactionUseCase.kt
@@ -46,15 +46,6 @@ class CreateFlexaTransactionUseCase(
                         )
                 }
 
-                is SubmitResult.GrpcFailure -> {
-                    Flexa
-                        .buildSpend()
-                        .transactionSent(
-                            commerceSessionId = transaction.getOrNull()?.commerceSessionId.orEmpty(),
-                            txSignature = result.txIds.first()
-                        )
-                }
-
                 else -> {
                     // do nothing
                 }