+++ ui-lib/src/main/java/co/electriccoin/zcash/ui/common/usecase/SendEmailUseCase.kt
@@ -119,11 +119,31 @@ class SendEmailUseCase(
      * Sends a support email for gRPC failure.
      */
     operator fun invoke(submitResult: SubmitResult.GrpcFailure) {
+        val reportDescription =
+            when (submitResult.reason) {
+                SubmitResult.GrpcFailure.Reason.TIMEOUT -> {
+                    context.getString(R.string.send_confirmation_pending_timeout_subtitle)
+                }
+
+                null -> {
+                    submitResult.description
+                }
+            }
+
         sendSupportEmail(
             subject = context.getString(R.string.app_name),
             messageBody =
                 EmailUtil.formatMessage(
-                    body = "Grpc failure",
+                    body =
+                        buildString {
+                            appendLine("Grpc failure")
+                            reportDescription
+                                ?.takeIf { it.isNotBlank() }
+                                ?.let {
+                                    appendLine()
+                                    appendLine(it)
+                                }
+                        },
                     supportInfo = ""
                 )
         )